pyproject.toml
[tool.poetry]
name = "parse-telegram"
version = "0.1.0"
description = ""
authors = ["Karamashev Leonid <leonkaram@mail.ru>"]
readme = "README.md"
packages = [{include = "parse_telegram"}]

[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.29.2"
python-dotenv = "^1.0.0"
aiohttp = "^3.8.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
